module: only allow white listed modules to be loaded
The kernel will generate a SHA1 hash of each module that is to be
loaded and check this hash against a list of known hashes. If a hash
matching that of the module-to-be-loaded is not found in this list,
the kernel will refuse to load the module.

The list of known hashes is passed into the kernel via
/sys/kernel/module_whitelist. The whitelist is a write-once list; after
the list is loaded it cannot be modified until the device is rebooted.
At the present time the list is limited to PAGE_SIZE (usually 4096
bytes).

Change-Id: If2959485ff6d743d745e0a3f60548699b4bfd026
Signed-off-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/518767
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Christopher Fries <qcf001@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
(cherry picked from commit c15985e40b630c72368089ccd5344e79ad8da3b0)

Conflicts:
	init/Kconfig
	kernel/Makefile