usb: dwc3: Allow rapid charging when usb is secure
If /sys/class/android_usb/android0/secure is set (when the device is
security locked), usb enumeration is disabled, hence limiting charging
to 100mA which does not effectively charge the device. Allow the device
to charge at bMaxPower as soon as SDP_CHARGER detection is complete, by
notifying the controller/otg driver when the secure attribute is set.
Remove unconditonal charging when the secure attribute is cleared.

Change-Id: I6591f3d9fbd43a684443dead163248e99a2d8937
Signed-off-by: Ravi Vembu <a18394@motorola.com>

Conflicts:
	drivers/usb/dwc3/dwc3_otg.h
	drivers/usb/gadget/android.c